### 2020-02-07 #2081 @Kevin-Mattheus-Moerman
@albansouche thanks for your submission! I'll be handling this submission. We'll first review the work for scope and if we deem it suitable for JOSS I'll work to find reviewers.
In the meantime, you could troubleshoot that missing DOI listed above. You can run `@whedon generate pdf` to update the proof here yourself.
