+++ paper/paper.bib
@@ -6,7 +6,7 @@
     Number = {1},
     Pages = {539 -- 578},
     Volume = {30},
-    Doi = {},
+    Doi = {10.1146/annurev.fluid.30.1.539},
     publisher = {}
 }
 
@@ -18,7 +18,7 @@
     Number = {4},
     Pages = {489 -- 490},
     Volume = {37},
-    Doi = {},
+    Doi = {10.1023/A:1020843529530},
     publisher = {}
 }
 
@@ -30,7 +30,7 @@
     Number = {24},
     Pages = {3039 -- 3067},
     Volume = {190},
-    Doi = {},
+    Doi = {10.1016/S0045-7825(00)00381-9},
     publisher = {}
 }
 
@@ -68,7 +68,7 @@
     Number = {89},
     Pages = {1456 -- 1467},
     Volume = {13},
-    Doi = {},
+    Doi = {10.1016/j.compstruc.2011.02.019},
     publisher = {}
 }
 
@@ -78,5 +78,6 @@
     Title = {Automated solution of differential equations by the finite element method: The FEniCS book.},
     Publisher = {Springer Science & Business Media},
     Volume = {84},
-  	Year = {2012}
+  	Year = {2012},
+    Doi = {10.1007/978-3-642-23099-8}
 }
