### 2020-02-08 #2081 @Kevin-Mattheus-Moerman
@chrisrichardson @garth-wells @chennachaos @weymouth @BenjaminRueth @JaroslavHron
Would you be interested in reviewing this work for [JOSS](joss.theoj.org/): **turtleFSI: A Robust and Monolithic FEniCS-based Fluid-Structure Interaction Solver**? The review focuses on [the software](https://github.com/KVSlab/turtleFSI.git) and [this short paper](https://github.com/openjournals/joss-papers/blob/joss.02081/joss.02081/10.21105.joss.02081.pdf)
Let me know if you have any questions.
