### 2020-02-10 #2081 @albansouche
@Kevin-Mattheus-Moerman Thank you for handling our submission. I have added the DOIs of the references in the new proof.
