+++ turtleFSI/problems/MMS_fluid.py
@@ -1,0 +1,187 @@
+# File under GNU GPL (v3) licence, see LICENSE file for details.
+# This software is distributed WITHOUT ANY WARRANTY; without even
+# the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+# PURPOSE.
+
+"""Problem file for running the method of manufactured solution for the fluid
+part of the equations. The manufactured solution is adapted from J. L. Guermond et al. 2005
+"""
+
+from dolfin import *
+import numpy as np
+from os import path
+
+from turtleFSI.problems import *
+from turtleFSI.modules import *
+
+
+def set_problem_parameters(default_variables, **namespace):
+    # Overwrite or add new variables to 'default_variables'
+    default_variables.update(dict(
+        # Temporal variables
+        T=0.00005,                # End time [s]
+        dt=0.00001,               # Time step [s]
+        theta=0.50001,            # Temporal scheme
+
+        # Physical constants
+        rho_f=1.0,                # Fluid density [kg/m3]
+        mu_f=1.0,                 # Fluid dynamic viscosity [Pa.s]
+
+        # Problem specific
+        folder="MMS_fluid",       # Name of the results folder
+        solid="no_solid",         # Do not solve for the solid
+        extrapolation="no_extrapolation",  # No displacement to extrapolate
+
+        # Geometric variables
+        N=40,                     # Mesh resolution
+
+        # MMS, From J. L. Guermond et al. 2005
+        eps = 1e-4,
+        #ux_mms = "pi * exp(3*t_e) * sin(2 * x[1]*x[1]) * sin(x[0]) * sin(x[0]) + eps + x[1] / pi + t_e + pow(t_e, 2)",
+        #uy_mms = "- pi * exp(3*t_e) * sin(2 * x[0]*x[0]) * sin(x[1]) * sin(x[1]) + eps + t_e + pow(t_e, 2)",
+        #p_mms = "exp(3*t_e) * cos(x[0]) * sin(x[1]) + eps + t_e + pow(t_e, 2)"))
+        ux_mms = "exp(3*t_e) + t_e + t_e * t_e * t_e + sin(t_e) + x[1] + x[0]",
+        uy_mms = "exp(3*t_e) + t_e + t_e * t_e * t_e + sin(t_e) - x[1] - x[0]",
+        p_mms = "3*exp(3*t_e) + t_e + t_e * t_e * t_e + sin(t_e)"))
+
+    return default_variables
+
+
+def get_mesh_domain_and_boundaries(N, **namespace):
+    # Load and refine mesh
+    mesh = RectangleMesh(Point(0, 0), Point(np.pi, np.pi), N, N)
+
+    # Mark the boundaries
+    Allboundaries = DomainBoundary()
+    boundaries = MeshFunction("size_t", mesh, mesh.geometry().dim() - 1)
+    boundaries.set_all(1)
+
+    # Define the domain
+    domains = MeshFunction("size_t", mesh, mesh.geometry().dim())
+    domains.set_all(1)
+
+    return mesh, domains, boundaries
+
+
+def initiate(F_fluid_nonlinear, theta, ux_mms, uy_mms, p_mms, dt,
+             v_deg, p_deg, eps, mesh, mu_f, psi, dx_f, dvp_, **namespace):
+    # Exact solution expressions
+    ux_e = Expression(ux_mms, eps=eps, degree=v_deg, t_e=0)
+    uy_e = Expression(uy_mms, eps=eps, degree=v_deg, t_e=0)
+    p_e = Expression(p_mms, eps=eps, degree=p_deg, t_e=0)
+    t_e_n = Constant(0.0)
[…]
