### 2020-04-01 #2089 @chennachaos
Hi @albansouche
I must say that it is very refreshing to see an open-source software library for FSI problems.

I have started the review. I have been going through the code and the associated documentation. I am having difficulties in understanding some parts. I would like to understand them at length before providing my detailed remarks.

1.) How, and on what level, does turtleFSI interact with FEniCS?

2.) What is the example problem used in the tests?

3.) Can the solver cope with mismatching meshes at the fluid-solid interface?

4.) Do you have any data on the convergence of Newton-Raphson iterations and how this convergence is affected under different settings of parameters passed to the *newtonsolver*?

5.) a) How is the Jacobian matrix for the monolithic scheme computed?
     b) How is the matrix system resulting from the monolithic scheme solved?

These (5a and 5b) are the two most time-consuming parts in an FSI solver. To gain insights into the computational aspects, I have looked at the graphs you have nicely prepared in the solver performance section. From the graphs, I notice that *one assembly of the Jacobian matrix requires about two orders of magnitude more CPU time than that of one matrix solver*. My observation seems to be in line with your statement "Notice that most of the compute time is spent on the Jacobian operations" in the caption.

This behaviour of the solver, however, is counter-intuitive to me because the matrix solver should be the most expensive part in any implicit scheme. Please clarify why the formation of the Jacobian matrix is so much more expensive.
