### 2020-04-02 #2089 @albansouche
Hi @chennachaos
Thank you for your first feedbacks. Here are my replies.

_1.) How, and on what level, does turtleFSI interact with FeniCS?_
turtleFSI interacts with FEniCS through the python interface. It provides a compact code relatively easy to modify for further development for anyone with basic knowledge of the FEniCS library.

_2.) What is the example problem used in the tests?_
We check all the examples (Turek flag with only solid, only fluid, and FSI) and all the mesh lifting operators in the tests.

_3.) Can the solver cope with mismatching meshes at the fluid-solid interface?_
Unfortunately not. turtleFSI solver only copes with conforming solid and fluid meshes at the interface.

_4.) Do you have any data on the convergence of Newton-Raphson iterations and how this convergence is affected under different settings of parameters passed to the newtonsolver?_
We didn’t document this aspect in details since the convergence rate can vary greatly depending on the problem being solved, the mesh resolution, or the time step…
In brief, the convergence is almost systematical quadratic when the Jacobian matrix is re-evaluated at each iteration. However, for efficiency reason, most of the time we prefer to reuse the Jacobian matrix over the Newton iterations even if this results in a sub-linear convergence rate. Recomputing the Jacobian matrix is not excessively time consuming by itself, but will  systematically require a new factorization step within the direct solver (here, MUMPS) which is computationally very expensive.

_5.) a) How is the Jacobian matrix for the monolithic scheme computed?_
The Jacobian matrix is numerically evaluated within FEniCS.

_b) How is the matrix system resulting from the monolithic scheme solved?_
It is solved with MUMPS solver available though the FEniCS PETSc backend. It appears to be the most robust solver available from FEniCS for FSI problems.

_These (5a and 5b) are the two most time-consuming parts in an FSI solver. To gain insights into the computational aspects, I have looked at the graphs you have nicely prepared in the solver performance section. From the graphs, I notice that one assembly of the Jacobian matrix requires about two orders of magnitude more CPU time than that of one matrix solver. My observation seems to be in line with your statement "Notice that most of the compute time is spent on the Jacobian operations" in the caption.
This behaviour of the solver, however, is counter-intuitive to me because the matrix solver should be the most expensive part in any implicit scheme. Please clarify why the formation of the Jacobian matrix is so much more expensive._

This is a very good observation and we also investigated this point further since the submission. Indeed the graphs presented in documentation (Solver performance) contain a mistake and will be rectified. The reported time for the Jacobian matrix evaluation and assembly was taken together with the preprocessing step of the MUMPS solver, which is  obviously a mistake. The preprocessing step in MUMPS corresponds to the factorization of the stiffness matrix of the linear system of equations. As to be expected, the factorization is the most memory and time consuming step for the direct solver. As mentioned in reply 4.), the computation and assembly of the Jacobian matrix is not the limiting factor regarding computational time, but will systematically require a new factorization within the direct solver which is computationally very expensive (roughly 2 orders of magnitude in comparison to the solving of the equations). There might be some options to interact with MUMPS directly through PETSc to control the re-factorization of the matrix for a given sparsity pattern. We have not investigated this part further at the moment.
