### 2020-04-08 #2089 @JaroslavHron
Hi @albansouche , your turtleFSI is really nice tool to easily setup and run  FSI problems.

- I was missing some info on the used time discretization. In the setup you supply parameter `theta=0.5` which probably means Crank-Nicolson method, but in FSI example you use `theta=0.51` - ia it some way to stabilize the computation - and is it sill 2nd order? May be I just missed it or It can be found in the referenced theses.

- Running the provided benchmark setups, I have experienced one problem when running them in parallel - it seems that the point evaluation at his point [TF_fsi.py#L205](https://github.com/KVSlab/turtleFSI/blob/5c5c33e90a0f75a2048290182228a435e1295766/turtleFSI/problems/TF_fsi.py#L205) fails in parallel.
As far as I know it is a limitation in fenics, but there are some ways to overcome this limitation using mpi4py (see.
https://fenicsproject.discourse.group/t/problem-with-evaluation-at-a-point-in-parallel)

- The only minor remark to the text I have: It might be useful to add the reference to the "Turek flag benchmarks" as is done in the documentation (https://turtlefsi2.readthedocs.io/en/latest/acknow_ref.html#id3) (but consider it as very minor remark - as I am coauthor of that publication and I am happy that it hase become such a popular test in part of the FSI community, however not everybody is familiar with that setup)
