### 2020-04-17 9da2a9e
fix figure_hpc + fix TF parallel + comments
