+++ paper/paper.bib
@@ -81,3 +81,12 @@
   	Year = {2012},
     Doi = {10.1007/978-3-642-23099-8}
 }
+
+@incollection{Turek:2006,
+  Title={Proposal for numerical benchmarking of fluid-structure interaction between an elastic object and laminar incompressible flow.},
+  author={{Turek}, Stefan and {Hron}, Jaroslav},
+  booktitle={Fluid-structure interaction},
+  pages={371--385},
+  year={2006},
+  publisher={Springer}
+}
