+++ docs/source/verif_perf.rst
@@ -22,6 +22,12 @@ A direct consequence of splitting the geometry in several MPI domains (Figure b)
 the system size associated with the handling of the degree of freedoms along the
 inner split boundaries. We illustrate this effect in Figure d where the total memory usage
 is monitored as function of the number of MPI tasks used to solve the problem.
+In the simulation, most of the time is spent in the factorization step within the direct solver (MUMPS)
+every time a new Jacobian matrix is evaluated. Therefore, reusing the Jacobian matrix over
+several Newton's iterations can be an efficient strategy if the solution shows a good convergence rate.
+In our example, we reuse the Jacobian matrix and the factorization of the direct solver over 5 Newton's iterations.
+As shown in Figure c, the total time in computing once the Jacobian matrix and
+the factorisation of the system is about 2 orders of magnitude larger than solving 5 iteration steps by reusing the factorisation.
 
 
 .. figure:: ../../figs/figure_HPC.png
@@ -30,9 +36,8 @@ is monitored as function of the number of MPI tasks used to solve the problem.
 
     Figure: Strong scaling of a 3D FSI pipe flow problem. a) Meshing of the inner fluid domain,
     and outer solid pipe with a total of ca. 63000 elements. b) Split of the geometry in
-    16 MPI domains. c) Time of computing and assembling the Jacobian matrix once (jac(1)) and
-    solving 5 Newton's iterations (it(5)) over one time step,
-    as function of the number of MPI tasks. Notice that most of the compute time is spent
-    on the Jacobian operations. Results obtained with a mesh discretization of 63000 elements.
+    16 MPI domains. c) Total time spent for one time step (jac.: jacobian matrix evaluation,
+    fact.: direct solver factorization step, it.: direct solver solve steps) as function of the number of MPI tasks.
+    Results obtained with a mesh discretization of 63000 elements.
     d) System memory usage as function of the number of MPI tasks for 3 different mesh discretizations
     of the problem illustrated in a).
