### 2020-04-17 8eba147
fix figure_HPC
