### 2020-04-17 #2089 @albansouche
Hi @JaroslavHron,
Thank you for the nice feedback and comments. Here are my replies:

Comment 1: Yes, theta=1 corresponds to a fully implicit scheme, theta=0 to a fully explicit scheme, and theta=0.5 to a Crank-Nicolson scheme. One may also use the so-called shifted Crank-Nicolson scheme that corresponds to 0.5 + dt (the time step), “which allows for global stability of the solution” (Wick, 2011) and preserves the convergence properties of the Crank-Nicolson scheme. In the examples we use a shifted Crank-Nicolson scheme with value of 0.51 when dt is set to 0.01 s.
You can find the reference here: Wick, T., 2011. Adaptive finite element simulation of fluid-structure interaction with application to heart-valve dynamics (Doctoral dissertation). https://doi.org/10.11588/heidok.00012992

- Comment 2: Thank you for pointing this out and it is, indeed, a FEniCS limitation. The issue arises when evaluating the displacement field at a given mesh point in the _post_solve_ function of the problem file. The user can modify this file as they want without impacting turtleFSI solver. So, this is not related to turtleFSI solver. However, since it might be of interest for the future users to execute the solver in parallel, we now added the fix in the examples TF_fsi.py and TF_csm.py.

- Comment 3:  Yes, we now added the reference in the text as well.
