### 2020-04-17 37aacb8
Turek paper doi
