+++ paper/paper.bib
@@ -88,5 +88,6 @@
   booktitle={Fluid-structure interaction},
   pages={371--385},
   year={2006},
+  Doi = {10.1007/3-540-34596-5_15},
   publisher={Springer}
 }
