### 2020-04-17 0080562
fix figure_hpc2
