### 2020-04-17 #2089 @albansouche
Hi @chennachaos,
Following your comment on the computational cost of the different parts of the solver, I modified some of the text and the figure present in the documentation (https://turtlefsi2.readthedocs.io/en/latest/verif_perf.html).
