### 2020-04-26 #2089 @chennachaos
Thanks for your response, @albansouche.

I have the following comments regarding the software paper.

* **Summary:** is nicely written.

* **A statement of need:** It is not clear what the motivation is and who the target audience are. Although the monolithic FSI solvers are robust, they are notorious to be computationally expensive beyond 100,000 DOFs, even for benchmark problems. Please note that benchmark problems are not always simple; one requires to run a considerable number of simulations (i) to assess the convergence of overall scheme, (ii) how different ingredients affect the accuracy, convergence and robustness, and (iii) to evaluate the parallel performance of the software tool.

* **State of the field:** This aspect is missing from the paper. There is no discussion and comparison with other FSI solvers (codes) available today. Oomph-lib and Elmer FEM are two such full-fledged open-source solvers. PreCICE is an open-source coupling library for FSI problems. Why should one prefer this FSI solver over others?

* **Quality of writing:** The paper is well written.

* **References:** The list of references needs to be improved by adding some classical references in the field of FSI.
