### 2020-04-26 #2089 @chennachaos
I still have some issues in understanding the computational times reported in the "Solver performance" page.
In a tetrahedral mesh with 63,000 first-order elements, there would be about 15000 nodes. Based on this, I estimate the total DOFs for the monolithic scheme to be about 60,000. My experience indicates that the matrix factorisation step should not take 1000 seconds (~16 minutes) (assuming that the optimisation flags are turned ON) for a matrix of this size.
I tried to understand this further from the graph, but it is not clear to me what "jac.(1) + fact.(1) + it.(5)" represents. Does it mean that you performed 5 iterations by computing and factorising the Jacobian matrix once? If this the case, then (a) how does the error norm of the residual converge and (b) what is the tolerance for convergence criterion?

I would like to see the computational time split into three parts: (i) for computing the Jacobian, (ii) for the factorisation and (iii) for the solver, if you reuse the Jacobian matrix, for the first 100 time steps. Also, please do provide the convergence of the residual norm in Newton-Raphson iterations for a few random time steps in the first 100 time steps.
