+++ turtleFSI/monolithic.py
@@ -53,7 +53,7 @@ vars().update(default_variables)
 vars().update(create_folders(**vars()))
 
 # Get mesh information
-mesh, domains, boundaries = get_mesh_domain_and_boundaries(**vars())
+mesh, domains, boundaries, constrained_domain = get_mesh_domain_and_boundaries(**vars())
 
 # Control FEniCS output
 set_log_level(loglevel)
@@ -74,7 +74,7 @@ if extrapolation == "biharmonic":
 else:
     Elem = MixedElement([de, ve, pe])
 
-DVP = FunctionSpace(mesh, Elem)
+DVP = FunctionSpace(mesh, Elem, constrained_domain=constrained_domain)
 
 # Create one function for time step n, n-1, and n-2
 dvp_ = {}
