+++ paper/paper.bib
@@ -73,21 +73,51 @@
 }
 
 @book{Logg:2012,
-  	Adsurl = {https://fenicsproject.org/book/},
-  	Author = {{Logg}, A. and {Mardal}, K-A. and {Wells}, G.},
+    Adsurl = {https://fenicsproject.org/book/},
+    Author = {{Logg}, A. and {Mardal}, K-A. and {Wells}, G.},
     Title = {Automated solution of differential equations by the finite element method: The FEniCS book.},
     Publisher = {Springer Science & Business Media},
     Volume = {84},
-  	Year = {2012},
+    Year = {2012},
     Doi = {10.1007/978-3-642-23099-8}
 }
 
 @incollection{Turek:2006,
-  Title={Proposal for numerical benchmarking of fluid-structure interaction between an elastic object and laminar incompressible flow.},
-  author={{Turek}, Stefan and {Hron}, Jaroslav},
-  booktitle={Fluid-structure interaction},
-  pages={371--385},
-  year={2006},
-  Doi = {10.1007/3-540-34596-5_15},
-  publisher={Springer}
+    Title={Proposal for numerical benchmarking of fluid-structure interaction between an elastic object and laminar incompressible flow.},
+    author={{Turek}, Stefan and {Hron}, Jaroslav},
+    booktitle={Fluid-structure interaction},
+    pages={371--385},
+    year={2006},
+    Doi = {10.1007/3-540-34596-5_15},
+    publisher={Springer}
 }
+
+@article{Malinen:2013,
+    title={Elmer finite element solver for multiphysics and multiscale problems},
+    author={Malinen, Mika and R{\aa}back, P},
+    journal={Multiscale Model. Methods Appl. Mater. Sci.},
+    volume={19},
+    pages={101--113},
+    year={2013},
+    publisher={Schriften des Forschungszentrums Julich, IAS Series Julich, Germany}
+}
+
+@incollection{Heil:2006,
+    title={oomph-lib--an object-oriented multi-physics finite-element library},
+    author={Heil, Matthias and Hazel, Andrew L},
+    booktitle={Fluid-structure interaction},
+    pages={19--49},
+    year={2006},
+    publisher={Springer}
+}
+
+@inproceedings{Jasak:2007,
+    title={OpenFOAM: A C++ library for complex physics simulations},
+    author={Jasak, Hrvoje and Jemcov, Aleksandar and Tukovic, Zeljko and others},
+    booktitle={International workshop on coupled methods in numerical dynamics},
+    volume={1000},
+    pages={1--20},
+    year={2007},
+    organization={IUC Dubrovnik Croatia}
+}
+
