### 2020-05-10 #2089 @aslakbergersen
@chennachaos, thank you for your comments. We apologize for the delayed response.

> Thanks for your response, @albansouche.
>
> I have the following comments regarding the software paper.
>
> * **Summary:** is nicely written.
> * **A statement of need:** It is not clear what the motivation is and who the target audience are. Although the monolithic FSI solvers are robust, they are notorious to be computationally expensive beyond 100,000 DOFs, even for benchmark problems. Please note that benchmark problems are not always simple; one requires to run a considerable number of simulations (i) to assess the convergence of overall scheme, (ii) how different ingredients affect the accuracy, convergence and robustness, and (iii) to evaluate the parallel performance of the software tool.
> * **State of the field:** This aspect is missing from the paper. There is no discussion and comparison with other FSI solvers (codes) available today. Oomph-lib and Elmer FEM are two such full-fledged open-source solvers. PreCICE is an open-source coupling library for FSI problems. Why should one prefer this FSI solver over others?
> * **Quality of writing:** The paper is well written.
> * **References:** The list of references needs to be improved by adding some classical references in the field of FSI.

To address the above points, we have added the text following below to the paper, with changes highlighted in bold. We have further specified the motivation, target audience, and why *some* could prefer turtleFSI over other solvers. We have also added some reference to other open-source solvers with FSI capabilities:

(…), but where physical insight might suffice. The aim was therefore to develop a fully monolithic and robust entry-level research code **with ease-of-use targeted towards students, educators, and researchers.**

FEniCS [@Logg:2012] has emerged as one (…) FEniCS was a natural choice of computing environment. **Compared to other open-source FSI solvers [@Malinen:2013; @Heil:2006; @Jasak:2007], turtleFSI is written in only a couple of hundred lines of high-level Python code, in contrast to tens of thousands of lines of low-level C++ code. This provides full transparency and a unique opportunity for researchers and educators to modify and experiment with the code, while still providing out of the box entry-level high-performance computing capabilities. Furthermore, because of the close resemblance between mathematics and code in FEniCS, users can make additions or modifications with ease.**

The turtleFSI solver rely on a fully monolithic (…) which can easily be adapted to other 2D or 3D FSI problems.

In conclusion, **turtleFSI is not a superior FSI solver in terms of speed, but it is a robust entry-level FSI solver** and performs exactly as designed and intended; ‘slow and steady wins the race’

----------------------------------------------

We have extended the section *Solver performance* from the documentation with a new section *Newton solver convergence and performance*. The documentation now includes a comparison of the Newton solver convergence between performing a full update and reusing the Jacobian matrix. This should answer the questions and concerns quoted below on the convergence behavior of the Newton solver used in different conditions. Of course, the reported behavior is not unique and depends on the geometry and boundary conditions but reflects well the general behavior we have observed using turtleFSI in our research work.

>I still have some issues in understanding the computational times reported in the "Solver performance" page. In a tetrahedral mesh with 63,000 first-order elements, there would be about 15000 nodes. Based on this, I estimate the total DOFs for the monolithic scheme to be about 60,000.

Our 3D FSI pipe flow problem has 460,000 dofs when discretized with 63000 tetrahedral elements. The pressure is approximated with a continuous linear scalar field, but the velocity and the deformation are both solved with a continuous quadratic vector field for stability reason.

>My experience indicates that the matrix factorisation step should not take 1000 seconds (~16 minutes) (assuming that the optimisation flags are turned ON) for a matrix of this size.
I tried to understand this further from the graph, but it is not clear to me what "jac.(1) + fact.(1) + it.(5)" represents. Does it mean that you performed 5 iterations by computing and factorising the Jacobian matrix once?

Yes, we compute and factorize the Jacobian matrix once at the first iteration and reused it for the next iterations. The factorization of the Jacobian matrix (with size 460000*460000) takes indeed ca. 1000 seconds running on 1 core and 1 thread.

> If this the case, then (a) how does the error norm of the residual converge and (b) what is the tolerance for convergence criterion? I would like to see the computational time split into three parts: (i) for computing the Jacobian, (ii) for the factorisation and (iii) for the solver, if you reuse the Jacobian matrix, for the first 100 time steps. Also, please do provide the convergence of the residual norm in Newton-Raphson iterations for a few random time steps in the first 100 time steps.

We have performed additional runs and documented the solver convergence rate in the new version of the section “Newton solver convergence and performance” of the software documentation.
