### 2020-05-20 58d4f43
Fix invalid DOI format
