+++ paper/paper.bib
@@ -108,7 +108,7 @@
     booktitle={Fluid-structure interaction},
     pages={19--49},
     year={2006},
-    doi={https://doi.org/10.1007/3-540-34596-5_2},
+    doi={10.1007/3-540-34596-5_2},
     publisher={Springer}
 }
 
