### 2020-05-20 #2089 @Kevin-Mattheus-Moerman
@nathanquinlan, @chennachaos, @JaroslavHron thanks for your review efforts! We greatly appreciate your time and help!

@chennachaos It looks like the authors made a number of changes which look ready for you to review.

@JaroslavHron your last unticked box relates to `references`, I believe these have now been fixed. Can you check and if possible tick the last box? Thanks!

@nathanquinlan are you able to start with your part of the review process? Thanks!
