### 2020-05-31 #2089 @chennachaos
I have checked the installation process and found that the installation is quite involved especially if one does not have **anaconda** installed. (Conda requires about 4-5 GB disk space).

However, I have used the development version and installed it by following the instructions in the documentation page. Everything worked without any issues. I was also able to run the default test case; I had to kill the job though, as the default settings (3000 time steps) requires a lot of time to run. My suggestion is to modify the default settings for the test cases so that the overall simulation time for the test cases is limited to a few minutes, perhaps with an accompanying note to the user.

It is fantastic that you included the well-known benchmark FSI example from the Turek and Hron paper. However, test cases with the time-periodic response of the structure require a significant amount of time to get some meaningful results beyond the onset of vortex-shedding.
To make it easy for the novice users to understand the code and run test cases, I suggest including a numerical example with a steady-state response, for example, a beam in crossflow, as studied in https://www.sciencedirect.com/science/article/pii/S0045782518301026
