### 2020-06-12 d9a992e
Update installation.rst
Update install instructions
