### 2020-06-12 #2089 @aslakbergersen
Hi @chennachaos, I apologize for the delay. See below for answers to your comments.

> I have checked the installation process and found that the installation is quite involved especially if one does not have anaconda installed. (Conda requires about 4-5 GB disk space).

Thank you for your diligence. We have added a note that recommends users to install miniconda over the full anaconda, and provide the following [link](https://docs.conda.io/projects/conda/en/latest/user-guide/install/download.html#anaconda-or-miniconda) for users to learn more about the difference.

> However, I have used the development version and installed it by following the instructions in the documentation page. Everything worked without any issues. I was also able to run the default test case; I had to kill the job though, as the default settings (3000 time steps) requires a lot of time to run. My suggestion is to modify the default settings for the test cases so that the overall simulation time for the test cases is limited to a few minutes, perhaps with an accompanying note to the user.
>
> It is fantastic that you included the well-known benchmark FSI example from the Turek and Hron paper. However, test cases with the time-periodic response of the structure require a significant amount of time to get some meaningful results beyond the onset of vortex-shedding.
To make it easy for the novice users to understand the code and run test cases, I suggest including a numerical example with a steady-state response, for example, a beam in crossflow, as studied in https://www.sciencedirect.com/science/article/pii/S0045782518301026

This is an excellent point we as developers did not think about. The `turtle_demo.py` setup was initially designed to simulate a turtle swimming over many cycles to check the stability of the setup and create the animation for the repository. We have now reduced the total time of the simulation of the turtle demo problem to execute only the first 30 time steps, which should run in a couple of minutes on a desktop computer. We have also added a note in the documentation to explain this further.
