### 2020-06-12 #2089 @chennachaos
HI @Kevin-Mattheus-Moerman. I now have completed my review.

The authors have addressed all the issues and improved the quality of code, tests and documentation to a satisfactory level.
I recommend the **turtleFSI** library for publication in the JOSS.

Thank you very much for the opportunity!
