### 2020-06-15 #2089 @nathanquinlan
Dear all,
Sorry I haven't been able to contribute to the review process to date.  I don't want to delay the process further so I have no objection if the editors would like to proceed without my review.
Regards
Nathan
