### 2020-06-15 #2089 @Kevin-Mattheus-Moerman
@albansouche I will process your paper for acceptance.

Here are some comments on your paper:
- [ ] Replace `rely` with `relies` in: `The turtleFSI solver rely on a fully monolithic approach...`
- [ ] in this sentence `We implemented and evaluated four different mesh lifting operators, ranging from simple and efficient 2nd order Laplace equation, most suitable for small deformations, to more sophisticated and computationally expensive 4th order bi-harmonic equation that can handle larger mesh deformations.`, it looks like this needs either `a simple` and `a more` (to go with the singular `equation`) or `equation` should be plural as in `equations`, please check.
- [ ] Minor typo in `...although the later is...`, this should have `latter`.
- [ ] Remove the `Installation and Use` section, this should be in the documentation and should therefore not appear in the paper.
- [ ] Add s to illustration in: `Here are some illustration of the...` i.e. use `illustrations`
- [ ] I would recommend increasing the size of the image in the paper, e.g. full width.
- [ ] Your paper is about to be accepted in JOSS, please thoroughly review/reread the paper yourself and check for typos etc. In particular check that the author names, affiliations, and acknowledgements are accurate.

Once you've made the above changes you can run `@whedon generate pdf` here to update the paper.

Finally please also do the following:
- [ ] Please archive a copy of the software on Zenodo and report back to us with the DOI of the archived version. Ensure that the archived version meta data on Zenodo matches the paper here. E.g. the author list and the title should match the JOSS paper. You can edit this after creating the archived version if needed.
- [ ] Can you confirm what the version of the final reviewed and archived version is? Is it still `v1.3`?
