### 2020-06-16 93480f4
Update size of pictures and typos
