### 2020-06-16 989d5fe
Update version number in setup.py
