### 2020-06-16 #2089 @aslakbergersen
@Kevin-Mattheus-Moerman thank you for your comments. I have implemented all your comments and checked for typos.

> Please archive a copy of the software on Zenodo and report back to us with the DOI of the archived version. Ensure that the archived version meta data on Zenodo matches the paper here. E.g. the author list and the title should match the JOSS paper. You can edit this after creating the archived version if needed.
10.5281/zenodo.3895443

> [Can you confirm what the version of the final reviewed and archived version is? Is it still v1.3?
To include the changes in the review to the distributed version on conda-forge I bumped the version number to 1.4. Do I need to edit this somewhere else? The DOI refers to the v1.4
